Adding type selector to McMd::DihedralDist

+++ src/mcMd/analyzers/base/DistributionAnalyzer.tpp
@@ -39,7 +39,7 @@ namespace McMd
       readOutputFileName(in);
       ParamComposite::template read<double>(in, "min", min_);
       ParamComposite::template read<double>(in, "max", max_);
-      ParamComposite::template read<double>(in, "nBin", nBin_);
+      ParamComposite::template read<int>(in, "nBin", nBin_);
       accumulator_.setParam(min_, max_, nBin_);
       accumulator_.clear();
    }
@@ -55,7 +55,7 @@ namespace McMd
       loadOutputFileName(ar);
       ParamComposite::template loadParameter<double>(ar, "min", min_);
       ParamComposite::template loadParameter<double>(ar, "max", max_);
-      ParamComposite::template loadParameter<double>(ar, "nBin", nBin_);
+      ParamComposite::template loadParameter<int>(ar, "nBin", nBin_);
       ar & accumulator_;
 
       // Validate
